move projects to project store
