+++ src/pages/IndexPage.vue
@@ -30,6 +30,7 @@ import { mapState } from "pinia";
 import AddProject from "../components/AddProject.vue";
 import Project from "../components/Project.vue";
 import { useAppStore } from "../stores/app";
+import { useProjectStore } from "../stores/project";
 
 export default defineComponent({
     name: "IndexPage",
@@ -38,11 +39,11 @@ export default defineComponent({
         Project
     },
     computed: {
-        ...mapState(useAppStore, ["getProjects"])
+        ...mapState(useProjectStore, ["getProjects"])
     },
     mounted() {
         useAppStore().setTitle("BranchWise");
-        useAppStore().setCurrentProject(null);
+        useProjectStore().setCurrentProject(null);
     }
 });
 </script>
